latest build

+++ test/testers/gemm-ukernel.h
@@ -88,7 +88,7 @@ public:
 	}
 
 	void testSGEMM(nnp_fast_sgemm_function fast_sgemm) const {
-		const uint_fast32_t seed = std::chrono::system_clock::now().time_since_epoch().count();
+		const uint_fast32_t seed = uint_fast32_t(std::chrono::system_clock::now().time_since_epoch().count());
 		auto rng = std::bind(std::uniform_real_distribution<float>(), std::mt19937(seed));
 
 		std::vector<float, AlignedAllocator<float, 32>> a(mr() * kc());
@@ -124,7 +124,7 @@ public:
 	}
 
 	void testSGEMM(nnp_full_sgemm_function full_sgemm) const {
-		const uint_fast32_t seed = std::chrono::system_clock::now().time_since_epoch().count();
+		const uint_fast32_t seed = uint_fast32_t(std::chrono::system_clock::now().time_since_epoch().count());
 		auto rng = std::bind(std::uniform_real_distribution<float>(), std::mt19937(seed));
 
 		std::vector<float, AlignedAllocator<float, 32>> a(mr() * kc());
